:ambulance: fix: Reconfigure jest to acknowledge ESM
