+++ package.json
@@ -2,14 +2,14 @@
   "name": "@namchee/telepon",
   "version": "1.0.7",
   "description": "Sane and easy-to-use Indonesian phone number validator and formatter.",
-  "exports": "./build/index.js",
-  "types": "./build/index.d.ts",
+  "exports": "./dist/index.js",
+  "types": "dist/index.d.ts",
   "type": "module",
   "scripts": {
     "test": "jest",
     "build": "rm -rf build && tsc",
-    "lint": "eslint index.ts __tests__/**/*.ts",
-    "lint:fix": "eslint index.ts __tests__/**/*.ts --fix"
+    "lint": "eslint src/**/*.ts",
+    "lint:fix": "eslint src/**/*.ts --fix"
   },
   "repository": {
     "type": "git",
@@ -33,21 +33,21 @@
   "devDependencies": {
     "@auto-it/conventional-commits": "^10.30.0",
     "@auto-it/npm": "^10.20.6",
-    "@jest/types": "^27.0.6",
-    "@types/jest": "^26.0.20",
-    "@typescript-eslint/eslint-plugin": "^4.15.0",
-    "@typescript-eslint/parser": "^4.15.0",
+    "@jest/types": "^27.2.4",
+    "@types/jest": "^27.0.2",
+    "@typescript-eslint/eslint-plugin": "^4.32.0",
+    "@typescript-eslint/parser": "^4.32.0",
     "auto": "^10.20.6",
     "cross-env": "^7.0.2",
     "dotenv": "^8.2.0",
     "eslint": "^7.19.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-jsdoc": "^31.6.1",
-    "jest": "^26.6.3",
+    "jest": "^27.2.4",
     "rimraf": "^3.0.2",
-    "ts-jest": "^26.5.3",
-    "ts-node": "^8.10.2",
-    "typescript": "^4.2.3",
+    "ts-jest": "^27.0.5",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.3",
     "typesync": "^0.8.0"
   },
   "publishConfig": {
