+++ tsconfig.json
@@ -5,9 +5,8 @@
       "ES2019.Array",
       "ES2020.Promise"
     ],
-    "module": "ES2015",
-    "allowJs": true,
-    "outDir": "./build",
+    "module": "ESNext",
+    "outDir": "./dist",
     "preserveConstEnums": true,
     "removeComments": true,
     "declaration": true,
@@ -27,6 +26,7 @@
     "index.ts",
   ],
   "exclude": [
+    "src/**/*.test.ts",
     "node_modules"
   ]
 }
